# docs/swagger.yaml
basePath: /api/v1
definitions:
  controller.BaseAndCodeAreEqualError:
    properties:
      message:
        type: string
    type: object
  controller.InvalidUUIDError:
    type: object
  controller.UnavailableCurrencyError:
    properties:
      currencyList:
        type: string
    type: object
  currencyrates.NoUUIDFoundError:
    properties:
      message:
        type: string
    type: object
  models.CurrencyRate:
    properties:
      base:
        type: string
      currency:
        type: string
      updated_at:
        type: string
      value:
        type: number
    type: object
  models.CurrencyUpdateUUID:
    properties:
      base:
        type: string
      currency:
        type: string
      uuid:
        type: string
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: localhost:8080
info:
  contact:
    email: valentinakuzmenkova@gmail.com
    name: Valentina Kuzmenkova
  description: 'Base currency is USD. The following currencies are supported: USD,
    EUR, MXN. Info is provided by VAT comply.'
  title: Swagger Currency Rates API
  version: "1.0"
paths:
  /rates:
    get:
      parameters:
      - description: uuid of update
        format: string
        in: query
        name: uuid
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CurrencyRate'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/currencyrates.NoUUIDFoundError'
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Gets rate by UUID
      tags:
      - rate
  /rates/{code}:
    get:
      parameters:
      - description: currency base, f.e. EUR
        format: string
        in: path
        name: code
        type: string
      - description: currency base, default=USD
        format: string
        in: query
        name: base
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CurrencyRate'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.BaseAndCodeAreEqualError'
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get the latest currency rate from the database
      tags:
      - rate
  /rates/{code}/update:
    put:
      parameters:
      - description: currency base, f.e. EUR
        format: string
        in: path
        name: code
        type: string
      - description: currency base, default=USD
        format: string
        in: query
        name: base
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CurrencyUpdateUUID'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.BaseAndCodeAreEqualError'
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Initiates an exchange rate update and provides a UUID of the update
      tags:
      - rate
swagger: "2.0"